% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian-Quadrature.R
\name{GenerateGaussLaguerreR}
\alias{GenerateGaussLaguerreR}
\title{Generate Gauss-Laguerre Quadrature in R}
\usage{
GenerateGaussLaguerreR(n)
}
\arguments{
\item{n}{}
}
\description{
Generate Gauss-Laguerre Quadrature in R
}
\examples{
GenerateGaussLaguerreR(5)
}
