% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian-Quadrature.R
\name{GenerateGaussLegendre}
\alias{GenerateGaussLegendre}
\title{Generate Gauss-Legendre Quadrature in R}
\usage{
GenerateGaussLegendre(n)
}
\arguments{
\item{n}{}
}
\description{
Generate Gauss-Legendre Quadrature in R
}
\examples{
GenerateGaussLegendre(5)
}
