% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston-Prob.R
\name{HestonProb}
\alias{HestonProb}
\title{Real Part from Heston Integrand}
\usage{
HestonProb(phi, kappa, theta, lambda, rho, sigma, tau, K, S, r, q, v0,
  Pnum)
}
\arguments{
\item{phi}{Integral variable}

\item{kappa}{Volatility mean reversion speed}

\item{theta}{Volatility mean reversion level}

\item{lambda}{risk parameter}

\item{rho}{correlation between 2 Brownian/Wiener Process}

\item{sigma}{volatility of variance in CIR pde}

\item{tau}{time to maturity}

\item{K}{strike price}

\item{S}{current stock price}

\item{r}{risk free rate}

\item{q}{dividend rate}

\item{v0}{initial value for variance}

\item{Pnum}{1 or 2 for P1 or P2 under risk-neutral measure}
}
\value{
real part of Heston integrand
}
\description{
Real Part from Heston Integrand
}
