% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiniteDifference.R
\name{MCCentral}
\alias{MCCentral}
\title{Title}
\usage{
MCCentral(S0, K, mu, sigma, time, paths, ds = 1e-05)
}
\arguments{
\item{ds}{}
}
\description{
Title
}
\examples{
MCCentral(100,120,0.01,0.25,1,1000)
}
