% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiniteDifference.R
\name{MCForward}
\alias{MCForward}
\title{Title}
\usage{
MCForward(S0, K, mu, sigma, time, paths, ds = 1e-05)
}
\arguments{
\item{ds}{}
}
\description{
Title
}
\examples{
MCForward(100,100,0.02,0.25,1,100)
}
