% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton-Cotes.R
\name{MidPoint}
\alias{MidPoint}
\title{Mid-point Rule for Integration}
\usage{
MidPoint(N, a, b)
}
\arguments{
\item{b}{}
}
\description{
Mid-point Rule for Integration
}
\examples{
MidPoint(N=1000,a=0.0001,b=100)
}
