% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBMSim.R
\name{SimGBM}
\alias{SimGBM}
\title{Single Path Stock Price Simulation for Geometric Brownian Motion}
\usage{
SimGBM(S0, mu, sigma, time, steps, type)
}
\arguments{
\item{S0}{Initial Stock Price}

\item{mu}{Annualized return in decimal format}

\item{sigma}{Annualized standard deviation in decimal format}

\item{time}{Time index (eg: 1,2,3,etc)}

\item{steps}{steps per Time Index (eg: 100, 200)}

\item{type}{"Continuous or Discrete"}
}
\value{
a vector of stock price based
}
\description{
Single Path Stock Price Simulation for Geometric Brownian Motion
}
\examples{
SimGBM(100,0.03,0.25,2,100,"discrete")
}
