% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBMSim.R
\name{SimTGBM}
\alias{SimTGBM}
\title{Multiple Paths Terminal Stock Price Simulation for Geometric Brownian Motion}
\usage{
SimTGBM(S0, mu, sigma, time, steps, paths, type)
}
\arguments{
\item{S0}{Initial Stock Price}

\item{mu}{Annualized return in decimal format}

\item{sigma}{Annualized standard deviation in decimal format}

\item{time}{Time index (eg: 1,2,3,etc)}

\item{steps}{steps per Time Index (eg: 100, 200)}

\item{paths}{Number of Paths in Simulation}

\item{type}{"Continuous or Discrete"}
}
\description{
Multiple Paths Terminal Stock Price Simulation for Geometric Brownian Motion
}
\examples{
SimTGBM(100,0.03,0.25,2,100,10,"continuous")
}
