% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton-Cotes.R
\name{Trapz}
\alias{Trapz}
\title{Trapezoidal Rule for Integration}
\usage{
Trapz(N, a, b)
}
\arguments{
\item{b}{}
}
\description{
Trapezoidal Rule for Integration
}
\examples{
Trapz(N=1000,a=0.0001,b=100)
}
